Added support for p4 shelve
Syntax: git p4 shelve [-d] [-c changelist]

Shelves all commits of the current branch that are not submitted to Perforce into a single changelist.

The parameter -c allows to specify the number of an existing changelist. All shelved files in the changelist will be replaced.
The parameter -d allows deleting an existing changelist with its shelved files.

Change-Id: I885d9b8fb1c58ca0ddc4271ff141051b9b2835b3